Rename namespace to plural overall
