+++ src/Submissions/Api/Areas/Admin/Models/Biobanks/ResetPasswordEntityModel.cs
@@ -1,4 +1,4 @@
-namespace Biobanks.Submissions.Api.Areas.Admin.Models.Biobank;
+namespace Biobanks.Submissions.Api.Areas.Admin.Models.Biobanks;
 
 public class ResetPasswordEntityModel
 {
