+++ src/Submissions/Api/Areas/Admin/Views/Biobanks/BiobankAdmin.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Http.Extensions
-@model Biobanks.Submissions.Api.Areas.Admin.Models.Biobank.BiobankModel
+@model Biobanks.Submissions.Api.Areas.Admin.Models.Biobanks.BiobankModel
 @{
     ViewBag.Title = "Biobank Admin";
 }
