+++ src/Submissions/Api/Areas/Admin/Views/Biobanks/Biobanks.cshtml
@@ -1,4 +1,4 @@
-@model Biobanks.Submissions.Api.Areas.Admin.Models.Biobank.BiobanksModel;
+@model Biobanks.Submissions.Api.Areas.Admin.Models.Biobanks.BiobanksModel;
 
 @{
     ViewBag.Title = "ADAC Admin";
