+++ src/Submissions/Api/Areas/Admin/Views/Biobanks/_ModalResetPassword.cshtml
@@ -1,4 +1,4 @@
-@model Biobanks.Submissions.Api.Areas.Admin.Models.Biobank.ResetPasswordEntityModel;
+@model Biobanks.Submissions.Api.Areas.Admin.Models.Biobanks.ResetPasswordEntityModel;
 
 <div class="modal-dialog">
   <div class="modal-content">
